% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_query.R
\name{add_ade_value_labels}
\alias{add_ade_value_labels}
\title{Add labels to values}
\usage{
add_ade_value_labels(.data, conn, drop_codes = F)
}
\arguments{
\item{.data}{Data obtained from the ADE API}

\item{conn}{The connection that produced the data}

\item{drop_codes}{If true, the labels will replace the code values. If false, codes are kept.}
}
\value{
A data.frame enriched with labels for values
}
\description{
By default the ADE API returns value codes for all variables, this function
replaces them with more meaningful labels
}
