% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{add_filter}
\alias{add_filter}
\title{Add filter to API query}
\usage{
add_filter(conn, var, values)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}

\item{var}{The variable to apply value filters to}

\item{values}{A vector of values for var to take}
}
\description{
Filters API query to only return a subset of records
}
\examples{
conn <- connect_to_ade("POPES_SUB_004")
conn <- add_filter(conn, var = "sex", values = c("SEX1", "SEX2"))


}
