% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{add_filters}
\alias{add_filters}
\title{Add filters to API query}
\usage{
add_filters(conn, ...)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}

\item{...}{a named list of vectors, where the name is the variable the filter applies to, and the vector contains the filter values}
}
\description{
Filters API query to only return a subset of records
}
\examples{
dontrun{
conn <- connect_to_ade("POPES_SUB_004")
conn <- add_filters(conn, sex = c("SEX1", "SEX2"), age = c("AGE1539", "AGE4064"))
}

}
