% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{build_filter_string}
\alias{build_filter_string}
\title{Build filter string}
\usage{
build_filter_string(conn)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}
}
\description{
Turn given filters for an API query into a string to be used in the query URL
}
