% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{clean_ade_names}
\alias{clean_ade_names}
\title{Clean API variable names}
\usage{
clean_ade_names(text, resourceId, var_type = "id")
}
\arguments{
\item{text}{a vector of strings to clean}

\item{resourceId}{resourceId for the variable}

\item{var_type}{used to infer prefix, does it come from a concept, codelist, etc}
}
\description{
The Aotearoa Data Explorer refers to all variable metadata with names of the form
<prefix>_<variable_name>_<series_name>. This function extracts that core <variable_name>
value and returns it in lower case.
}
\examples{
clean_ade_names("CL_AREA_POPES_SUB_004", var_type = "codelist", resourceId = "POPES_SUB_004")
> "area"

}
