% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{collect.API_conn}
\alias{collect.API_conn}
\title{Collect data from API}
\usage{
\method{collect}{API_conn}(conn, add_labels = T, drop_codes = F)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}

\item{add_labels}{Add text labels for variables?}

\item{drop_codes}{If adding text labels, remove the code version of variables?}
}
\value{
A data.table object
}
\description{
Executes the API query generated for the series
}
