% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_query.R
\name{get_ade_metadata}
\alias{get_ade_metadata}
\title{Get metadata for resource}
\usage{
get_ade_metadata(resourceId, version, providerId)
}
\arguments{
\item{resourceId}{an API key generated at https://portal.apis.stats.govt.nz/}
}
\value{
A list of data.frame objects containing:
\itemize{
\item series - descriptors of the data set being queried
\item concepts - metadata on the variables included in the data
\item codelists - the values each variable can take, and a textual name for them
}
}
\description{
Request certain useful metadata values from the structural metadata API
Used by connect_to_ade()
}
\examples{
\dontrun{
get_ade_metadata("POPES_SUB_004", version = "1.0")
}
}
