% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_ade_query_header}
\alias{get_ade_query_header}
\title{Get header for use in Aotearoa Data Explorer API queries}
\usage{
get_ade_query_header(format = NULL)
}
\arguments{
\item{format}{optional parameter to specify return format for queries, defaults to XML}
}
\description{
Get header for use in Aotearoa Data Explorer API queries
}
\examples{
header <- get_ade_query_header(format = "text/csv")

}
