% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_series.R
\name{get_all_data_series}
\alias{get_all_data_series}
\title{Get all data series}
\usage{
get_all_data_series()
}
\description{
Searches the titles of all published datasets in the ADE and returns a data frame
containing textual name, and the flowRef and version required to query it
}
