% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_query.R
\name{get_series_metadata}
\alias{get_series_metadata}
\title{Get series metadata}
\usage{
get_series_metadata(conn)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}
}
\description{
Get the name, description, and other info about the series for the connection object
}
\examples{
\dontrun{
conn <- connect_to_ade("POPES_SUB_004")
series_info <- get_series_metadata(conn)
series_info$Name.en
> "STATSNZ Subnational population estimates (RC, SA2), by age and sex, at 30 June 1996-2024 (2024 boundaries)"
}
}
