% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_query.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Get variable code-label pairs}
\usage{
get_variable_labels(conn, variable)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}

\item{variable}{The name of a variable in the dataset for which the metadata applies}
}
\value{
A data.frame
}
\description{
Get the codes and respective labels for values a variable can take
}
\examples{
metadata <- get_ade_metadata("POPES_SUB_004")
age_groups <- get_variable_labels(conn, variable = "AGE_POPES_SUB_004")

}
