% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_query.R
\name{get_variable_values}
\alias{get_variable_values}
\title{Get code-label pairs for all values of a variable}
\usage{
get_variable_values(conn, variable)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}

\item{variable}{The name of a variable in the dataset for which the metadata applies}
}
\value{
A data.frame
}
\description{
Get the codes and respective labels for values a variable can take
}
\examples{
\dontrun{
conn <- connect_to_ade("POPES_SUB_004")
age_groups <- get_variable_labels(conn, variable = "age")
age_groups <- get_variable_values(conn, variable = "age")
}
}
