% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_query.R
\name{get_variables}
\alias{get_variables}
\title{Get variables}
\usage{
get_variables(conn)
}
\arguments{
\item{metadata}{A metadata table obtained from the get_ade_metadata() function}
}
\value{
A character vector of variable names
}
\description{
List variables available in a data series
}
\examples{
metadata <- get_ade_metadata("POPES_SUB_004")
age_groups <- get_variables(metadata)

}
