% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{print.API_conn}
\alias{print.API_conn}
\title{Print}
\usage{
\method{print}{API_conn}(conn)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}
}
\description{
Prints the name of the series of an API_conn object
}
