% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{remove_filter}
\alias{remove_filter}
\title{Remove filter from API query}
\usage{
remove_filter(conn, var)
}
\arguments{
\item{conn}{A connection to the ADE API, obtained from connect_to_ade()}

\item{var}{The name of a variable with a filter currently applied by add_filters()}
}
\description{
Removes a filter from API query to the ADE
}
\examples{
dontrun{
conn <- connect_to_ade("POPES_SUB_004")
conn <- add_filters(conn, sex = c("SEX1", "SEX2"), age = c("AGE1539", "AGE4064"))
data <- collect(conn)
conn <- remove_filter(conn, var = "sex")
}
}
