% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{remove_leading_area_code}
\alias{remove_leading_area_code}
\title{Get area type}
\usage{
remove_leading_area_code(area)
}
\arguments{
\item{area}{a character string containing a label/name for an area concept}
}
\description{
Labels for some area types will include the area code in the label, this function removes the code.
}
\details{
NOTE: If you wish to use the get_area_type() function, use it before removing area codes.
Area codes are used to help classify SA2s and Auckland Local Boards.
}
\examples{
remove_leading_area_code("100100 North Cape")
> "North Cape"

}
