% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_series.R
\name{search_data_series}
\alias{search_data_series}
\title{Search data series}
\usage{
search_data_series(text, perl = F)
}
\arguments{
\item{text}{substring to search for}

\item{perl}{use perl regex?}
}
\description{
Searches the titles of all published datasets in the ADE and returns matches
}
\examples{
search_data_series(text="population")

}
