% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_api_environment}
\alias{set_api_environment}
\title{Set API environment}
\usage{
set_api_environment(environment)
}
\arguments{
\item{environment}{prod or alpha}
}
\description{
API environment variable to determine which API endpoint to use
}
\examples{
set_environment("alpha")

}
