% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzapi-package.R
\docType{package}
\name{snzapi-package}
\alias{snzapi}
\alias{snzapi-package}
\title{snzapi: Stats NZ API}
\description{
Query the Stats NZ Aotearoa Data Explorer API A collection of functions to help users query the Stats NZ API Built upon the rsdmx package created by Emmanuel Blondel, Matthieu Stigler, and Eric Persson
}
\keyword{internal}
