% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_query.R
\name{connect_to_ade}
\alias{connect_to_ade}
\title{Set up connection to a data series in the ADE}
\usage{
connect_to_ade(resourceId, version = NULL, format = "csv")
}
\arguments{
\item{resourceId}{the resourceId for a data series.
This can be found in the data explorer, or using the search_data_series() or get_all_data_series() functions.}
}
\description{
Create an S4 object to act as the basis of an API query
}
\examples{
conn <- connect_to_ade("POPES_SUB_004")

}
