% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_area_type}
\alias{get_area_type}
\title{Get area type}
\usage{
get_area_type(
  area_name,
  area_code = NULL,
  options = c("SA1", "SA2", "Region", "TA", "LB", "CB", "Subdivision", "DHB",
    "Constituency", "Ward")
)
}
\arguments{
\item{area_name}{a character string containing a label/name for an area concept}

\item{area_code}{(optional) the numeric code for an area to help improve classification accuracy}

\item{options}{(optional) possible options of types for the area}
}
\description{
Some published series include multiple area types in the same "area" column.
This helper function aims to assign, as best as it can, the relevant area type.
If you are querying a table with only one area type, you are best to set the area
type manually.
}
\details{
Known issues:
- If no code is provided, Auckland is assigned TA, although it may be a DHB
- SA2s need either their code included in the area_name, or need their code provided
- Totals and areas outside common geographies are often mislabeled
- Doesn't work for Urban-Rural areas
}
\examples{
get_area_type("Gisborne district")
> "TA"

}
