% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_area_type}
\alias{get_area_type}
\title{Get area type}
\usage{
get_area_type(area)
}
\arguments{
\item{area}{a character string containing a label/name for an area concept}
}
\description{
Some published series include multiple area types in the same "area" column.
This helper function aims to assign, as best as it can, the relevant area type.
If you are querying a table with only one area type, you are best to set the area
type manually.
}
\details{
Known issues:
- Auckland is assigned TA, but is also the name of a DHB
- Totals and areas outside common geographies are often mislabeled
- Doesn't work for Urban-Rural areas
}
\examples{
get_area_type("Gisborne district")
> "TA"

}
